Changed content layout of library.html
The 'content div' of library.html now uses justify-evenly to center them in case of only one item in a row.
Changed the sizing of the items inside the 'content div' so that only one item is in a row in smaller screen sizes.
TailwindCSS is now used to replace more inline CSS.